useAction
use action cache,
as the config needs a functional component - needed to abstract it one layer further to allow for a single load instead of each time
