+++ apps/web/actions/index.js
@@ -1,13 +1,5 @@
-import { useConnections } from './useConnections';
-import { useInstructions } from './useInstructions';
-import { useStateActions } from './useStateActions';
-import { useTesters } from './useTesters';
+import { loadedActions } from 'core/actions';
 
 export const useActions = () => {
-  const connections = useConnections();
-  const testers = useTesters();
-  const stateActions = useStateActions();
-  const instructions = useInstructions();
-
-  return { ...connections, ...testers, ...stateActions, ...instructions };
+  return loadedActions;
 };
